Move around and consolidate some
