+++ controllers/api/conditionController.js
@@ -1,8 +1,13 @@
-const Equipment = require('../models/equipment');
+const Condition = require('../../models/condition');
 const utility = require('./utility');
 
 exports.index = (req, res, next) => {
-  Equipment.find((err, _data) => {
+  var search_queries = {};
+  if (req.query.name !== undefined) {
+    search_queries.name = req.query.name;
+  }
+
+  Condition.find(search_queries, (err, _data) => {
     if (err) {
       next(err);
     }
@@ -17,7 +22,7 @@ exports.index = (req, res, next) => {
 };
 
 exports.show = (req, res, next) => {
-  Equipment.findOne({ index: req.params.index }, (err, data) => {
+  Condition.findOne({ index: req.params.index }, (err, data) => {
     if (err) {
       next(err);
     }
