+++ controllers/api/equipmentCategoryController.js
@@ -1,13 +1,8 @@
-const DamageType = require('../models/damageType');
+const EquipmentCategory = require('../../models/equipmentCategory');
 const utility = require('./utility');
 
 exports.index = (req, res, next) => {
-  var search_queries = {};
-  if (req.query.name !== undefined) {
-    search_queries.name = req.query.name;
-  }
-
-  DamageType.find(search_queries, (err, _data) => {
+  EquipmentCategory.find((err, _data) => {
     if (err) {
       next(err);
     }
@@ -20,9 +15,8 @@ exports.index = (req, res, next) => {
       res.status(200).json(utility.NamedAPIResource(data));
     });
 };
-
 exports.show = (req, res, next) => {
-  DamageType.findOne({ index: req.params.index }, (err, data) => {
+  EquipmentCategory.findOne({ index: req.params.index }, (err, data) => {
     if (err) {
       next(err);
     }
