+++ controllers/api/languageController.js
@@ -1,13 +1,8 @@
-const WeaponProperty = require('../models/weaponProperty');
+const Language = require('../../models/language');
 const utility = require('./utility');
 
 exports.index = (req, res, next) => {
-  var search_queries = {};
-  if (req.query.name !== undefined) {
-    search_queries.name = req.query.name;
-  }
-
-  WeaponProperty.find(search_queries, (err, _data) => {
+  Language.find((err, _data) => {
     if (err) {
       next(err);
     }
@@ -22,7 +17,7 @@ exports.index = (req, res, next) => {
 };
 
 exports.show = (req, res, next) => {
-  WeaponProperty.findOne({ index: req.params.index }, (err, data) => {
+  Language.findOne({ index: req.params.index }, (err, data) => {
     if (err) {
       next(err);
     }
