+++ controllers/api/startingEquipmentController.js
@@ -1,8 +1,8 @@
-const Subrace = require('../models/subrace');
+const StartingEquipment = require('../../models/startingEquipment');
 const utility = require('./utility');
 
 exports.index = (req, res, next) => {
-  Subrace.find((err, _data) => {
+  StartingEquipment.find((err, _data) => {
     if (err) {
       next(err);
     }
@@ -12,28 +12,32 @@ exports.index = (req, res, next) => {
       if (err) {
         next(err);
       }
-      res.status(200).json(utility.NamedAPIResource(data));
+      res.status(200).json(utility.ClassAPIResource(data));
     });
 };
 
 exports.show = (req, res, next) => {
   // search by class
-  if (utility.isRaceName(req.params.index) === true) {
-    Subrace.find({ 'race.name': utility.race_map[req.params.index] }, (err, _data) => {
-      if (err) {
-        next(err);
+
+  if (utility.isClassName(req.params.index) === true) {
+    StartingEquipment.findOne(
+      { 'class.name': utility.upperFirst(req.params.index) },
+      (err, _data) => {
+        if (err) {
+          next(err);
+        }
       }
-    })
+    )
       .sort({ url: 'asc', level: 'asc' })
       .exec((err, data) => {
         if (err) {
           next(err);
         }
-        res.status(200).json(utility.NamedAPIResource(data));
+        res.status(200).json(data);
       });
   } else {
     // return specific document
-    Subrace.findOne({ index: req.params.index }, (err, data) => {
+    StartingEquipment.findOne({ index: parseInt(req.params.index) }, (err, data) => {
       if (err) {
         next(err);
       }
