+++ controllers/api/traitController.js
@@ -1,0 +1,62 @@
+const Trait = require('../../models/trait');
+const utility = require('./utility');
+
+exports.index = (req, res, next) => {
+  Trait.find((err, _data) => {
+    if (err) {
+      next(err);
+    }
+  })
+    .sort({ index: 'asc' })
+    .exec((err, data) => {
+      if (err) {
+        next(err);
+      }
+      res.status(200).json(utility.NamedAPIResource(data));
+    });
+};
+
+exports.show = (req, res, next) => {
+  // search by class
+
+  if (utility.isRaceName(req.params.index) === true) {
+    Trait.find({ 'races.name': utility.upperFirst(req.params.index) }, (err, _data) => {
+      if (err) {
+        next(err);
+      }
+    })
+      .sort({ url: 'asc' })
+      .exec((err, data) => {
+        if (err) {
+          next(err);
+        }
+        res.status(200).json(utility.NamedAPIResource(data));
+      });
+  } else if (utility.isSubraceName(req.params.index) === true) {
+    Trait.find({ 'races.name': utility.subrace_map[req.params.index] }, (err, _data) => {
+      if (err) {
+        next(err);
+      }
+    })
+      .sort({ url: 'asc' })
+      .exec((err, data) => {
+        if (err) {
+          next(err);
+        }
+        res.status(200).json(utility.NamedAPIResource(data));
+      });
+  } else {
+    // return specific document
+    Trait.findOne({ index: req.params.index }, (err, data) => {
+      if (err) {
+        next(err);
+      }
+
+      if (data) {
+        res.status(200).json(data);
+      } else {
+        res.status(404).json({ error: 'Not found' });
+      }
+    });
+  }
+};
