+++ routes/equipment.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const EquipmentController = require('../controllers/equipmentController');
+const EquipmentController = require('../controllers/api/equipmentController');
 
 router.get('/', EquipmentController.index);
 router.get('/:index', EquipmentController.show);
