+++ routes/equipmentCategories.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const EquipmentCategoryController = require('../controllers/equipmentCategoryController');
+const EquipmentCategoryController = require('../controllers/api/equipmentCategoryController');
 
 router.get('/', EquipmentCategoryController.index);
 router.get('/:index', EquipmentCategoryController.show);
