+++ routes/magicSchools.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const MagicSchoolController = require('../controllers/magicSchoolController');
+const MagicSchoolController = require('../controllers/api/magicSchoolController');
 
 router.get('/', MagicSchoolController.index);
 router.get('/:index', MagicSchoolController.show);
