+++ routes/monsters.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const MonsterController = require('../controllers/monsterController');
+const MonsterController = require('../controllers/api/monsterController');
 
 router.get('/', MonsterController.index);
 router.get('/:index', MonsterController.show);
