+++ routes/skills.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const SkillController = require('../controllers/skillController');
+const SkillController = require('../controllers/api/skillController');
 
 router.get('/', SkillController.index);
 router.get('/:index', SkillController.show);
