+++ routes/startingequipment.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const StartingEquipmentController = require('../controllers/startingEquipmentController');
+const StartingEquipmentController = require('../controllers/api/startingEquipmentController');
 
 router.get('/', StartingEquipmentController.index);
 router.get('/:index', StartingEquipmentController.show);
