+++ routes/subraces.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const SubraceController = require('../controllers/subraceController');
+const SubraceController = require('../controllers/api/subraceController');
 
 router.get('/', SubraceController.index);
 router.get('/:index', SubraceController.show);
