+++ routes/traits.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const TraitController = require('../controllers/traitController');
+const TraitController = require('../controllers/api/traitController');
 
 router.get('/', TraitController.index);
 router.get('/:index', TraitController.show);
