+++ routes/weaponProperties.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const WeaponPropertyController = require('../controllers/weaponPropertyController');
+const WeaponPropertyController = require('../controllers/api/weaponPropertyController');
 
 router.get('/', WeaponPropertyController.index);
 router.get('/:index', WeaponPropertyController.show);
